requirements.txt
streamlit>=1.24.0
deriv-api>=1.0.0
pandas>=1.3.0
numpy>=1.21.0
plotly>=5.3.0
python-dotenv>=0.19.0
uvicorn>=0.15.0
pydantic>=1.8.0
websockets>=10.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
